+++ src/modules/quick-reply/quick-reply.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { QuickReplyController } from './quick-reply.controller';
+import { QuickReplyService } from './quick-reply.service';
+import { SessionModule } from '../session/session.module';
+
+@Module({
+  imports: [SessionModule],
+  controllers: [QuickReplyController],
+  providers: [QuickReplyService],
+  exports: [QuickReplyService],
+})
+export class QuickReplyModule {}